requirements.txt
python-dotenv
openai
pandas
numpy
pyyml
flask
markdown